Updated wordle example.

+++ examples/wordle words/the_entire_wordle.js
@@ -1,16 +1,18 @@
 const df = require("../../").default;
 const fs = require("node:fs");
 
-const raw = fs.readFileSync("raw_wordle_list.txt"); // Add your own words if you want to.
-const theEntireWordle = raw.toString()
-  .split("\n") // Split the strings.
-  .filter(v => !!v) // Filter falsey strings, such as empty ones.
+let raw = fs.readFileSync("dictionary.json"); // Add your own words if you want to.
+if (String.fromCharCode(raw[0]) != "[") raw = raw.subarray(3);
+
+const theEntireWordle = JSON.parse(raw)
+  .filter(v => v.length === 5) // Filter falsey strings, such as empty ones.
   .sort(() => Math.random() >= 0.5 ? 1 : -1); // Shufle.
 
 const wordsAmount = process.argv[2] || theEntireWordle.length;
+if (wordsAmount > theEntireWordle.length) throw new Error(`There's only ${theEntireWordle.length} words`);
 
-if(fs.existsSync("wordle_list_serialized_templ.json")) fs.rmSync("wordle_list_serialized_templ.json");
-if(fs.existsSync("wordle_list_templ.txt")) fs.rmSync("wordle_list_templ.txt");
+if (fs.existsSync("wordle_list_serialized_templ.json")) fs.rmSync("wordle_list_serialized_templ.json");
+if (fs.existsSync("wordle_list_templ.txt")) fs.rmSync("wordle_list_templ.txt");
 
 const template = df("The entire wordle.", (e, s) => {
   s.author = "UserUNP";
@@ -20,7 +22,7 @@ const template = df("The entire wordle.", (e, s) => {
   const items = [];
   e.defAction("addWords", (words) => {
     const chunk = e.mc("minecraft:book", `wordle chunk ${items.length+1}`, 1);
-    for(const word of words) chunk.addLore(`${word}`);
+    chunk.addLore(`${words.join(",")}`);
     items.push(chunk);
     console.log(`finished chunk ${items.length} (${words.length})`);
   });
@@ -33,7 +35,7 @@ const template = df("The entire wordle.", (e, s) => {
       e.action.addWords(wordsBuffer);
       wordsBuffer = [];
     }
-    if(i == theEntireWordle.length-1) {
+    if (i == theEntireWordle.length-1) {
       e.action.addWords(wordsBuffer);
       break;
     }
@@ -41,12 +43,12 @@ const template = df("The entire wordle.", (e, s) => {
   }
 
   wordsBuffer = [];
-  for(let i = 0; i < items.length; i++) {
+  for (let i = 0; i < items.length; i++) {
     if(i!=0 && i % 26 == 0) {
       e.setVar("AppendValue", listVar, ...wordsBuffer);
       wordsBuffer = [];
     }
-    if(i == theEntireWordle.length-1) {
+    if (i == theEntireWordle.length-1) {
       e.setVar("AppendValue", listVar, ...wordsBuffer);
       break;
     }
